backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0,>=3.2.0
python-multipart
pydantic-settings
email-validator
alembic
pytest
redis
